refactor(internal): fsStreamWriter
